Add files via upload
Flight Data contain all the information of flights from source to target with the respective IATA codes for the cities they are travelling to with their country code. Airport.csv file has longitude and latitude for the countries with respective IATA codes as the places to travel. Flight Data has data of one month for different years divided into 2003-2009 and 2010-2016 sheets. Our objective is to plot network for only one month of any year we want. 

I have used data from the year 2003 and stored it in 2003.csv file. Cleaning and filtering of data is done in Excel sheet itself for the year 2003 and on Jupyter Code we have used filtering for USA, China, UK and Australia. 

Python libraries used for manipulating and visualizing graphs is networkx with matplotlib and manipulating geospatial data is cartopy. Apart from this Numpy and Pandas are used for data manipulation and cleaning.